+++ src/utils/read.ts
@@ -43,7 +43,7 @@ export const getEnvironment = async (options: {
 
 	try {
 		if (options.envString) {
-			return parseEnvironment(options.envString);
+			return parseJsonStringEnvironment(options.envString);
 		}
 
 		const environment = config({
@@ -64,115 +64,14 @@ export const getEnvironment = async (options: {
 };
 
 /**
- * Convert a stringified environment to `Environment` type
+ * Convert a stringified json environment to `Environment` type
  *
- * @param envString - Stringified environment
+ * @param envJsonString - Stringified json environment
  * @returns Environment
  */
-export const parseEnvironment = (envString: string): Environment => {
-	const environment: Environment = {};
-	const lines = envString.split('\n');
-	lines.forEach((line) => {
-		setEnvironment(line, environment);
-	});
+export const parseJsonStringEnvironment = (
+	envJsonString: string,
+): Environment => {
+	const environment: Environment = JSON.parse(envJsonString);
 	return environment;
-};
-
-/**
- * Find content that is between a character in a string
- *
- * @example const string = 'this is a test "banana", that has text!';
- *
- * const result = findContentBetween(string, '"');
- *
- * console.log(result); // 'banana'
- *
- * @param string - String to search
- * @param character - Character to search for
- * @param reverse - Search from the end of the string
- * @returns the content between the character, or undefined if not found.
- */
-export const findContentBetween = (
-	string: string,
-	character: string,
-	reverse = false,
-): string | undefined => {
-	const start = reverse
-		? getSecondToLastIndexOf(string, character)
-		: string.indexOf(character);
-	const end = reverse
-		? string.lastIndexOf(character)
-		: getSecondToFirstIndexOf(string, character);
-
-	if (start === -1 || end === -1) {
-		return;
-	}
-	return string.substring(start + 1, end).trim();
-};
-
-const setEnvironment = (line: string, environment: Environment): void => {
-	// get all instances of `=` in the line
-	const equalSignIndexes = getIndexesOf(line, '=');
-	if (equalSignIndexes.length === 0) {
-		return;
-	}
-	let lastIndex = 0;
-	for (const equalSignIndex of equalSignIndexes) {
-		const currentLine = line.substring(lastIndex, equalSignIndex);
-		setEnvironmentFromLine(currentLine, environment);
-		lastIndex = currentLine.length;
-	}
-};
-
-const getIndexesOf = (string: string, character: string): number[] => {
-	const indexes: number[] = [];
-	let index = string.indexOf(character);
-	while (index !== -1) {
-		indexes.push(index);
-		index = string.indexOf(character, index + 1);
-	}
-	return indexes;
-};
-
-const setEnvironmentFromLine = (
-	line: string,
-	environment: Environment,
-): void => {
-	const index = line.indexOf('=');
-	if (index === -1) {
-		return;
-	}
-	const lineBefore = line.substring(0, index);
-	const lineAfter = line.substring(index + 1);
-
-	const environmentKey = getEnvironmentKey(lineBefore);
-	const environmentValue = getEnvironmentValue(lineAfter);
-
-	if (!environmentKey || !environmentValue) {
-		return;
-	}
-
-	environment[environmentKey] = environmentValue;
-};
-
-const getEnvironmentKey = (line: string): string | undefined =>
-	findContentBetween(line, '"', true);
-
-const getEnvironmentValue = (line: string): string | undefined =>
-	findContentBetween(line, '"');
-
-const getSecondToLastIndexOf = (string: string, character: string): number => {
-	const index = string.lastIndexOf(character);
-	if (index === -1) {
-		return -1;
-	}
-	return string.lastIndexOf(character, index - 1);
-};
-
-const getSecondToFirstIndexOf = (string: string, character: string): number => {
-	const index = string.indexOf(character);
-	if (index === -1) {
-		return -1;
-	}
-	return string.indexOf(character, index + 1);
 };